requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-heroku==0.3.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
jws==0.1.3
oauth2client==4.1.3
Pillow==8.3.1
protobuf==3.18.0rc1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
PyJWT==2.1.0
pyparsing==2.4.7
Pyrebase4==4.5.0
python-jwt==2.0.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
